move DeaParameters to test and automate allowed_parameters
